CHROMIUM: ARM: tegra: seaboard: add new config parameters for atmel_mxt_ts
Atmel has made a change to their driver to allow new conifg parameters
to be specified.  Those parameters need to be in the mxt_config_data
variable or the driver will fail to load.

This CL imports the new values (translated from a config file provide
by Atmel) in preparation for the import of that CL.

Change-Id: I2bef316632593c99e98fb9404116237094b406c0

R=olofj@chromium.org
BUG=NONE
TEST=Built patch on its own and with future Atmel imports
  Validated boot and operation of this CL only
    and in conjunction with future changes

Review URL: http://codereview.chromium.org/6711011