msm: acpuclock-8960: Use for_each_present_cpu for L2 voting
In the hotplug insertion path, acpuclk_set_rate() gets called
before the CPU has been brought online to the scheduler. As a
result calls to for_each_online_cpu() in this path will not
include the CPU that's being brought online

compute_l2_speed() returns an incorrect value because of this.
Replacing the use of for_each_online_cpu() in that function
with with for_each_present_cpu() resolves this.

Change-Id: If9529c00876e6fe1b25ac4fc939faa07f3a491ad
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>