CHROMIUM: ARM use generic BUG() handler on ARM
ARM uses its own BUG() handler for no good reason that I can see. This makes
it slightly different from the x86 implementation (and most others). To
make things easier for crash_reporter this moves it over to use the same
mechanism.

One of the problems is that the ARM implementation doesn't report the function
with the BUG() in it, but always reports the PC being in __bug(). The generic
implementation doesn't have this problem.

BUG=chromium-os:12454
TEST=echo bug > /proc/breakme, check that BUG output looks OK.

Change-Id: I515db9a04e98084e6bbb21c4a1d13579568a0904

Review URL: http://codereview.chromium.org/6597027