Bluetooth: Add support for storing the key size
In some cases it will be useful having the key size used for
encrypting the link. For example, some profiles may restrict
some operations depending on the key length.

The key size is stored in the key that is passed to userspace
using the pin_length field in the key structure.

For now this field is only valid for LE controllers. 3.0+HS
controllers define the Read Encryption Key Size command, this
field is intended for storing the value returned by that
command.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@openbossa.org>
(cherry picked from commit 99e812cade86bdc0f7675d93d5ef7fb7727fd658)

Change-Id: I48c5567020105497ffbe4d810adb50797064d60a
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>