msm: ramdump: Add the ramdump driver
The ram dump driver allows dumping of arbitrary memory
regions to userspace. When there is no userspace
daemon ready to collect the ram dump, the driver is
intelligent enough to immediately return a failure
instead of waiting for a long timeout. This is very
useful in cases where the ram dump is blocking
recovery of an error condition that's visible to the
user.

Change-Id: I1f551c823bac5404b685d5e21c31d6bd71f18f3f
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>