CHROMIUM: video: tegra: Wire up output connectedness
This makes the core dc driver keep track of whether a particular input
in enabled.  It is up to the output ops to maintain the connected status
if a detect op is plugged in, otherwise it is assumed that the output is
always connected.

BUG=chrome-os-partner:1812
TEST=After the patch series, ran X server on Seaboard, verified HDMI hotplug works with xrandr
Signed-off-by: Robert Morell <rmorell@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6677087