CHROMIUM: video: tegra: Expose possible bound head mask
This change adds support for userspace to query which heads an output
may be bound to.  Since our implementation currently hardcodes the head
to output mapping, we currently always fill in a single bit.  However,
for future implementations we will be able to dynamically bind outputs
to heads.

BUG=chrome-os-partner:1812
TEST=After the patch series, ran X server on Seaboard, verified HDMI hotplug works with xrandr
Signed-off-by: Robert Morell <rmorell@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6708069