mfd: wcd9310-core: Initial WCD9310 Driver commit
WCD9310 is a hardware audio codec that has 10
TX ports and 7 RX ports for audio streams.
It is capable of using SLIMBUS or I2S to send
audio streams, and its register set can be
accessed through either SLIMBUS or I2C.

It has many internal interrupt lines which are ORed
into one interrupt line coming out of the codec.

Change-Id: Ia7c8faae95feca5cfb2726642c062a7e3166309d
Signed-off-by: Brad Rubin <brubin@codeaurora.org>