msm_xo: Add support for XO_CORE
The XO_CORE controls the XO feeding the TCXO buffers (A0, A1,
etc.). Some users of the XO buffers want to keep the XO core on
even though their buffer is off or in pin control because it can
take a long time turn the XO back on and warm up the buffers.
Expose an XO control for such users so they can get better
power/performance.

Change-Id: Ie7452c53e68d4fb417675714d64e4e7f0798edb0
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>