ASoC: WCD9310: Add initial WCD9310 Codec Driver
The WCD9310 Codec Driver is an ALSA-compliant codec
driver that calls the WCD9310 Core Driver to access
codec registers and interrupts. This driver constructs
the internal codec audio paths with DAPM widgets and
controls and provides the controls to the upper layers
to enable and configure audio paths. It also registers
the WCD9310 Codec with the ALSA core.

Change-Id: I346a76db9a53745af7157c5a58b937aac7486f8d
Signed-off-by: Brad Rubin <brubin@codeaurora.org>