msm: clock-debug: Implement measure rate generically
Don't use clock ops to measure the rate of a clock. Instead,
provide a measure clock that accepts any measurable clock as its
parent. Calling clk_set_parent() on the measurement clock with a
clock will program the measurement hardware to measure that clock
and calling clk_get_rate() on the measurement clock will return
the rate of the measurement clock's parent.

Change-Id: I09a60011e06650a64fc0263ab890e029917f413c
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>