LOCAL: build_all.py: allow alternate build targets.
Allow make targets to be given to build-all.py, which
in turn allows the build-all preverification step to be used to
generate all kernel artifacts for later preverification steps.
This moves us closer to eliminating the need to build the kernel
twice for each commit to be tested by the PV framework.

Change-Id: I1688e7c2e408667440e3e8e9b1de52d6fa40c5a8
Signed-off-by: Gregory Bean <gbean@codeaurora.org>