Bluetooth: Fix keeping the command timer running
In the teardown path the reset command is sent to the controller,
this event causes the command timer to be reactivated.

So the timer is removed in two situations, when the adapter isn't
marked as UP and when we know that some command has been sent.

Reported-by: Keith Packard <keithp@keithp.com>
Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@openbossa.org>
Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>
(cherry picked from commit b79f44c16a4e2181b1d6423afe746745d5e949ff)

Change-Id: I5756a0035f63fe0a85f00adb0e1071773d15f377
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>