msm: timer: reset local timer count only at first boot
Previously we were resetting this timer each time we booted. Sometimes
alarms based on the old timer count were being used, causing events to
not come in until the timer caught up to the old time. Fix this by not
explicitly resetting the timer except on the first time we boot.

Change-Id: I27503580cc275226d5cdf00d471f528b970c0dbf
CRs-Fixed: 281314
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>