CHROMIUM: tegra: seaboard: fix ULPI transceiver clock
On our seaboard-based platforms, pll_p_out4 is used to clock
(through DAP_MCLK2) the external USB ULPI transceiver between the USB2
port to mini-PCIe modem. And it needs to be set at 24Mhz for the
tranceiver to work correctly.
Revert the sclk parent to pll_c out1 (as it was before the last change)
since it needs to stay clocked at 108MHz.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:3167
TEST=Check that the Gobi modem is appearing in lsusb listing
Use the kaen DVT for browsing and suspending

Review URL: http://codereview.chromium.org/6825076

Change-Id: Ia9b1bbc7860da732fb8edbbcf0bce60e82d6d8ed