mtd: msm_nand: Handle ADM stall errors in nand driver
Memory allocated using dma_alloc_coherent is expected to be
unbuffered and uncached. Nevertheless if we define the option
CONFIG_ARM_DMA_MEM_BUFFERABLE, this memory can be buffered.

Any system that uses ARMv7 processors with the L2 cache on
AXI bus will be required to explicitly drain their write
buffers. When memory is marked as bufferable, L2 cache
allocates the write buffers to the memory area. Doing a DSB
will only take care of the CPU and L1 caches. The outer cache
(i.e. PL310) has no means of knowing that a DSB occurred
(there are no indications from the CPU outside).

Currently NAND driver using only the DSB calls before and after
the ADM operation to make sure that all the memory contents are
available to the ADM before it starts processing.

But now DSB alone is not sufficient to flush the L2xO write
buffers. Need to add explicit 'outer_sync' calls to drain the
L2xO write buffers after every DSB call in the code.

Cache sync should not add performance impact. Any system that
implements will need to do this. Even if one may not be doing
an explicit cache-sync, hardware will have an equivalent operation.

Without this patch we see ADM stall errors as the command pointer
is not properly updated into main memory before the ADM starts
processing it.

Change-Id: I1360b3c4e7012d7e02401e2a07006b7576847927
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>