mmc: configure SDIO slot GPIOs in low-power mode during suspend
As of today, the suspend for SDIO devices means just turning off clocks.
With this patch, GPIOs(except DATA1) will also be configured to low-power
mode during suspend and will be restored to active mode during resume.
But since we would need DATA1 line to be active to receive SDIO wake up
events, it will be always be in active state. This change is required to
reduce overall system current consumption.

CRs-fixed: 280575
Change-Id: I7fb1af500534a52b9a3f7709908192eabfb1447e
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>