msm_charger: Charge from USB if enumerated
The usb driver may call in to the charger driver asking it to start
charging from the USB path even when the USB charging driver is not
initialized.  Later when the USB charger driver initializes it
doesn't begin charging and waits for usb's request.

Fix this by remembering the charge current from usb driver and start
charging as soon as the charging driver initializes.

CRs-Fixed: 283347
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>