qup_i2c: Don't call complete() with a NULL completion object
The QUP I2C interrupt handler might call the complete() function with an
NULL completion object resulting in a kernel crash.  This can happen in
response to an unexpected/spurious interrupt.  The interrupt handler does
check for a valid completion object but there is a potential multi-core
timing hazard where the completion object might be set to NULL after the
object is checked and before the complete() function is called.

The driver needs to make sure that the completion object is valid while
interrupts are enabled.

Change-Id: I662ac234329a2dd4968ceab6d0467b3f9bc22c1b
CRs-fixed: 283330
Signed-off-by: Kenneth Heitke <kheitke@codeaurora.org>