CHROMIUM: video: tegra: Add hook for DDC enable
Some output types like HDMI need to take special action when performing
a DDC transaction such as an EDID read.  This adds hooks to give
platform code the opportunity to set up for DDC transactions.

For HDMI, DDC access is enabled before the debouncing delay is applied;
this is necessary for some HDMI sinks to give the +5V time to settle and
the i2c microcontroller in the sink (powered off the +5V from the cable)
time to power up.

TEST=Test HDMI hotplug with Seaboard and Kaen, both before and after enabling HDMI with X11.
Signed-off-by: Robert Morell <rmorell@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6877112