ASoc: 8x60: Add support to route PCM session to multiple COPP
When FM transmission is in progress, if any system sound or notification
comes, this sound needs to routed to FM tx device and speaker. With
current adm driver and audio_dev_ctl driver routing session to
multiple devices is not supported. Update adm driver and
audio_dev_ctl to support following cases:
	1. Single session routing to multiple devices
	2. Multi Session routing to single device

Change-Id: I540f31c7ec6afb442202cfbd1ed54d6f33b23fda
CRs-fixed: 280808
Signed-off-by: Deepa Madiregama <dmadireg@codeaurora.org>