CHROMIUM: gobi: Remove worker.exit hack.
The exit member was added to hack around crosbug.com/14190, but we need a proper
fix. I discussed with olofj, and we think the right thing is to use a waitqueue
and wait for there to be work available or for the thread to be asked to exit.

BUG=chromium-os:14205
TEST=Adhoc
Build, modem status, suspend/resume

Signed-off-by: Elly Jones <ellyjones@chromium.org>
Signed-off-by: Jason Glasgow <jglasgow@chromium.org>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6883145

Change-Id: I8eaeb95c5c1a4426768f632c5144f7d5356e3a3a
(cherry picked from commit c69b798ef6849b97f23b0b9047592193b99f13af)
Reviewed-on: http://gerrit.chromium.org/gerrit/597
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: Elly Jones <ellyjones@chromium.org>