msm: acpuclock-8960: Don't change CPU clock source for SETRATE_HOTPLUG
When responding to a CPU_DEAD or CPU_UP_PREPARE notifier,
acpuclock cannot change CPU clock muxes for the dead CPU. This
is because the CP15 registers required to do so are per-CPU and
only accessible from that CPU, which is already offline.

Therefore skip changing and restoring the clock CPU clock mux
state for reason == SETRATE_HOTPLUG.  Continuing to turn off
the CPU's PLL (if is was still on) should be fine, however.

Change-Id: I965795004369f446b06d697a50292b31cf3fefa3
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>