CHROMIUM: ARM: Tegra: Seaboard: Add vdd_dmic regulator consumer
LDO-6 supplies power to both various portions of the camera circuitry,
and the digital microphones on the camera board.

Use cases exist where the microphones are used without the camera.
If only the camera driver controls LDO-6, the microphones will not
be powered in this case. To solve this, add an extra consumer of
LDO-6, which the audio driver will control. The regulator driver will
then turn on LDO-6 if either or both the audio and camera drivers are
active.

BUG=chrome-os-partner:3240
TEST=Tested recording from DMIC:
  Run alsamixer to configure system for DMIC capture:
  alsamixer
    Left Capture Mux: Left
    Left ADC Input: DMIC
    Right ADC Input: Toggle to ADC then back to DMIC
  Record audio:
    arecord -r 48000 -c 2 -f s16 -V stereo foo.wav
  Play audio:
    aplay foo.wav
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6895005