msm: board-msm8x60: set regulators 8901_l4 and l6 to be always on for SURFs
In order for additional power savings to take place, the RPM sets
various settings to be lower on the PMIC 8901 if all of its
regulators are in LPM during sleep.  This is a problem for SURFs
that have not received a rework for the XO_OUT_D1_EN pin because
they will not be able to resume after the PMIC 8901 has been put
into this state.  Therefore, ensure that some of the PMIC 8901
regulators are left always on in HPM for SURFs only.

Change-Id: I0a7569ecca5d55bed21e14c5f16bc05d4cf8256f
Signed-off-by: David Collins <collinsd@codeaurora.org>