msm: 8660: audio: Fix problem with Lineout right bias
There is an error in the Timpani register profiles for
speaker and speaker + headphone where the Lineout right
bias is not given time to settle before an audio
stream begins. This can cause a slightly louder pop
when enabling either speaker or speaker + headphone.

Also fixes duplicate macro problem in Timpani profiles
file.

Change-Id: I49040b5a83834366528ac7896befcd4cc771ad6f
Signed-off-by: Brad Rubin <brubin@codeaurora.org>