i2c-qup: For multiple write transfers, fill output FIFO as much as possible
When a transaction has multiple write transfers, fill the output FIFO with
as much data as possible rather than just the data for a single transfer.

If the total message length is greater than the output FIFO size, make
sure that the clock state is RESET_BUSIDLE or FORCED_LOW before entering
the RESET state.  Transitioning to the RESET state will cause the clock
state to be reset as well which can cause the i2c core to incorrectly
terminate the pending i2c transaction.

CRs-fixed: 282678
Change-Id: If1da9441c1fa7b7bc881efd42929f7c540977681
Signed-off-by: Kenneth Heitke <kheitke@codeaurora.org>