msm: gpio-v2: Convert gpio from a platform device to a sysdev device
The sysdev suspend and resume callbacks are executed after the I bit
is disabled in CPSR. The current platform suspend and resume methods
do not allow for a correct suspend/resume behavior since they would
unmask a wakeup interrupt too early in the suspend path. This would
allow a wakeup interrupt to be fired after we disable all irqs
and hence be masked before we go into suspend.
Once this happens, we won't be woken up by that interrupt which is bad.

Change-Id: I6f1a5e4201843c054e50f9d59e2add9ff867fb04
CRs-Fixed: 280718
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>