CHROMIUM: verity: remove compare_hash layer
We've refactored the code such that we are now only calling compare_hash
into two places in the same function. So we really don't need this layer
any more.

I plan to send a CL soonish that will remove the second call so we'll
be down to just the one call.

BUG=chromium-os:13872
TEST=Ran dm-verity.git unit tests. Ran platform_DMVerityCorruption on H/W.

Change-Id: Ieb20dfdcf425a3c423ebabf63f86d549c259c0f8
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>

R=wad@chromium.org,taysom@chromium.org

Review URL: http://codereview.chromium.org/6902084