video:msm: Initial QSEED sharpening/smoothing API
Adds an additional parameter (.dpp.sharp_strength) and reuses flag
MDP_SHARPENING to mdp_overlay, to allow for overlay_set to request for
sharpening/smoothing.

The overlay parameter is designed to range between -127 and 127, with
negative values infering smoothing, positive infering sharpening. This
range of values is then mapped onto 7 levels for both smoothing and
sharpening (i.e. ~ every increase/decrease of 16 aligns with an increased
sharpening level).

This commit only adds the API for sharpening, a patch that updates the
QSEED table2 values is needed for the desired effect to be visible.

Change-Id: Ib6522e43bde4f4a54e752258553822e78a95749d
Signed-off-by: Carl Vanderlip <carlv@codeaurora.org>