GIC: mask private interrupts in suspend
The suspend function assumes that the private interrupts even if not
marked wakeup don't need to be masked as they will never trigger while
suspended.
However that assumption is not true certain timer interrupts (ex DGT)
wraps up and causes an interrupt. This interrupt will cause a spurious
wakeup from suspend.

Fix the suspend function such that private interrupts are not treated
specially and will be masked if not marked IRQ_WAKEUP.

Change-Id: Ia53648bf473b2797183188a8e73f9a77051713ad
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>