CHROMIUM: arm: driver for accessing shared memory, first cut.
This change borrows from the previously submitted for review
http://codereview.chromium.org/6880063 (it was considered
too impacting) and needs to be accompanied by the config
change as per http://codereview.chromium.org/6902094

Also, for this to work it is required  that u-boot creates
a certain memory structure at 1M below the top of the
available physical memory, and modifies the kernel command
line to exclude the top 1 megabyte of DRAM from the kernel's
control.

It was envisioned originally, that drivers/platform/x86/chromeos.c
would be a common module providing dm-verity with interface
to platform specific modules, but as of now that file is too
much x86 centric, so a new interface file,
drivers/platform/arm/chromeos.c is introduced with this CL,
these two files will have to be merged with the underlying
drivers refactored.

Apart form adding new config options, this CL introduces the
actual driver which discovers the shared memory object and
exposes the shared memory as a file in /sys/kernel/debug to
be parsed by crossystem.

Blending in of the actual gpio states into the mapped
memory, as well as handling of nonvolatile storage will be
added in the upcoming CLs.

Change-Id: Ia961081ac1a7f760b4fa67b691c2774d810f3369
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>

BUG=chromium-os:12522
TEST=manual

- recompile the kernel
- scp chromeos_arm.ko to the target's /var directory
- on the target:
  insmod /var/chromeos_arm.ko
  rmmod chromeos_arm
  insmod /var/chromeos_arm.ko
  dmesg | grep chromeos_arm:
- observe messages indicating installing/removal of
  the driver
- observe that the file /sys/kernel/debug/chromeos_arm gets
  created/removed with driver installation/removal

localhost var # wc -c /sys/kernel/debug/chromeos_arm
892 /sys/kernel/debug/chromeos_arm

Review URL: http://codereview.chromium.org/6883204