msm: audio: qdsp5v2: Updated PCM out driver to allocate memory during init
pmem for the pcm out driver is allocated when the first audio playback
request comes to it. This is causing memory partition in Video playback
use cases as pcm driver retains the memory till next reboot.
Moving audio driver to late_init and pmem alloc is now done in init
to avoid partitioning of memory.

CRs-Fixed: 282400
Change-Id: Ib5b5f6651df731b6fcacfc7f152753339c6e2bd6
Signed-off-by: Phani Kumar Allada <pallad@codeaurora.org>