staging: ath6kl: open/shared auth implementation
If the wpa_supplicant conf file supplies both open and shared
algorithm, and AP is configured as shared then connection never
happens. Since it is a FMAC driver additional logic is added in
driver which first detects this, then tries open algorithm for the
first time and when it fails tries the shared algo.

kvalo: fix style issues

Signed-off-by: Naveen Singh <nsingh@atheros.com>
Signed-off-by: Kalle Valo <kalle.valo@atheros.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

BUG=chromium-os-partner:2296
TEST=SecMat (shared key auth tests)

Change-Id: I281336ab05e1c4c6256864febda96d7e70e89afc
Reviewed-on: http://gerrit.chromium.org/gerrit/446
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>