serial: tegra_hsuart: Use resources instead of platform data
Use resources to pass data to tegra_hsuart to set
mapbase, irq.

This change ports a previous change that was in the 37 kernel
but failed to port cleanly on the 38 kernel.

From: Jaikumar Ganesh <jaikumar@google.com>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>

BUG=chrome-os-partner:1949
TEST=if driver enabled without this change, it causes null pointer
derefernce trying to parse the membase field from the platform
data. With this change, the driver no longer panics.

Change-Id: Ie918ee1cb2baf774d56dbb2d29e0a23094080f1a
Reviewed-on: http://gerrit.chromium.org/gerrit/358
Tested-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>