msm: board-msm8960-regulator: remove pin ctrl column from regulator table
Remove the pin control parameter in the msm_pm8921_regulator_pdata
because it is no longer used for normal regulators.  It only needs to
be passed into the initialization data of pin control regulators.

Also, change the allowed operating modes to reflect changes to the
pm8921-regulator driver.  Only REGULATOR_MODE_NORMAL and
REGULATOR_MODE_IDLE can be used now.  Further limit regulators 8921_s5
and 8921_s6 to only REGULATOR_MODE_NORMAL because these regulators
supply the Krait cores and are not allowed to be in LPM.

Change-Id: I125915c44c10ceee50af07198b7df350238f7c0a
Signed-off-by: David Collins <collinsd@codeaurora.org>