CHROMIUM: media: tegra: Added more output formats.
All YUYV order variants supported for both input and output.
I420 (YUV planar) output also now supported.

In the process of this, there was some nontrivial code refactoring and cleanup.
Kind of hard to separate, because the additional format support relies on this
refactoring.

Signed-off-by: Andrew Chew <achew@nvidia.com>

BUG=none
TEST=Used the audiovideo_V4L2 test to capture at YUYV, YVYU, UYVY, VYUY, and
YU12. YU12 doesn't look correct right now because of a bug in the
testbench when doing the conversion between YU12 and RGB. I used a
separate program to do the conversion, and the images look fine.

Reviewed-on: http://gerrit.chromium.org/gerrit/394
Tested-by: Jie Sun <jiesun@chromium.org>
Reviewed-by: Jie Sun <jiesun@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>

Change-Id: Ia72ef430e8d3a6e98fdf88785b2d10e2111e6de7
Reviewed-on: https://gerrit.chromium.org/gerrit/11965
Commit-Ready: Pawel Osciak <posciak@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>