msm: kgsl: Delay 30ms to make sure GPU is ready during resume time
During resume, the kgsl driver preforms a soft reset of the GPU and it
needs to wait certain amount of time for GPU to be ready.
On most chips 10ms is sufficient. But on 8660 with newer RPM images,
the GPU often hangs with the 10ms delay, but testing has shown that
30ms does not hang.

CRs-Fixed: 286145
Change-Id: I85734bae383ba57c86af2c35e35ac4c3e2048a96
Signed-off-by: Wei Zou <wzou@codeaurora.org>