CHROMIUM: arm/dt: full tegra devicetree support
This patch adds full probing of devices from device tree data bu adding
a new board-dt.c board support file for device tree matching.

tegra250.dtsi describes everything within the Tegra SoC, but leaves all
optional devices disabled.

tegra-harmony.dts is the configuration of the board. It enables the optional
devices that are required for the board, and provides any required
configuration data for those devices.

Thanks to Stephen Warren for a bunch of testing and bug fixes to this patch.

Signed-off-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>