CHROMIUM: gobi: Remove worker.exit hack.
The exit member was added to hack around crosbug.com/14190, but we need a proper
fix. I discussed with olofj, and we think the right thing is to use a waitqueue
and wait for there to be work available or for the thread to be asked to exit.

BUG=chromium-os:14205
TEST=Adhoc
Build, modem status, suspend/resume

Signed-off-by: Elly Jones <ellyjones@chromium.org>
Signed-off-by: Jason Glasgow <jglasgow@chromium.org>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Change-Id: Ic1a990d6ab71861031a7c9149df3dcfef3b11111
Committed: http://chrome-svn/viewvc/chromeos?view=rev&revision=c69b798

Review URL: http://codereview.chromium.org/6913008