CHROMIUM: video: tegra: Add an extra queue for hotplug
Calling cancel_delayed_work() in an interrupt context may deadlock, if
the interrupt is handled while the workqueue is being processed.  This
is protected with a WARN_ON since commit 466bd303097.

To avoid this, add another layer of queueing that schedules the debounce
work.  This new layer of queuing is executed in a context where it is
possible to call cancel_delayed_work_sync().

Additionally, this adds a bit of cleanup and documentation along the
way.

(It is not desirable to simply change the existing call to the
non-interruptiple __cancel_delayed_work(), since if the irq is handled
while the work handler is running, the work handler may call
disable_ddc() and leave DDC disabled during the debounce interval.)

Signed-off-by: Robert Morell <rmorell@nvidia.com>

TEST=Hotplug HDMI on Seaboard, verify display is detected and kernel
emits not warnings

Change-Id: I66e2ffdd04e80d1558f943130d6457bdc2953194
Reviewed-on: http://gerrit.chromium.org/gerrit/152
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>