HACK: net: lock around SIOCGIFNAME operation
Grab the rtnl lock around dev_ifname so SIOCGIFNAME is interlocked
against device registration work. This is a gross hack to deal with
a race in device registration whereby udev and RTNL events are sent
up before the device is registered. This breaks flimflam who immediately
turns around and requests the device name with SIOCGIFNAME.

Change-Id: I83fd596181f67bf1e4a82bba9075956fe9854df4
Signed-off-by: <sleffler@chromium.org>

BUG=chromium-os:13159
TEST=boot on multiple devices and verify wlan0 exists

Change-Id: Iefaba08ad9d417eac572a43cf600c385a5a242a6
Reviewed-on: http://gerrit.chromium.org/gerrit/179
Tested-by: Micah Catlin <micahc@chromium.org>
Reviewed-by: Micah Catlin <micahc@chromium.org>