CHROMIUM: staging: ath6kl: mark nl80211 support for separate default keys
The WPA-PSK association was not going through with 2.6.38
kernel. It was found that supplicant was not able to set
the PTK key. On further analysis it was found that the function
nl80211_set_key in file net/wireless/nl80211.c is returning with
code as -EOPNOTSUPP. This function has been modified to check for
the flag "WIPHY_FLAG_SUPPORTS_SEPARATE_DEFAULT_KEYS"in wiphy
struct. If this flag is not set in the driver init, it returns
from here thereby causing the association to fail. The solution
is to set this flag in driver init as there would be separate
default keys for unicast and multicast packets.

Signed-off-by: Naveen Singh <nsingh@atheros.com>

BUG=none
TEST=SecMat

Change-Id: I75ccbd9dc1bbe31c39602ab5a1e459ce63cf6efe
Reviewed-on: http://gerrit.chromium.org/gerrit/180
Tested-by: Micah Catlin <micahc@chromium.org>
Reviewed-by: Micah Catlin <micahc@chromium.org>