msm: clock: Fix error path refcounting
When a branch clock fails to enable the power rail, we disable
the branch and return an error, but we forgot to undo setting the
enable flag. Similarly, if the refcounts are off and a user calls
clk_disable() twice, the count for a clock will underflow and
wrap around to UINT_MAX. In this situation, WARN_ON() and don't
do anything.

Change-Id: I39da7b05a030815350e40ac7454db9c3f3d63cad
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>