Pull Elly's Change from R12 in to 0.11.241.B Factory Branch Manually:
CHROMIUM: gobi: Remove worker.exit hack.

The exit member was added to hack around crosbug.com/14190, but we need a proper
fix. I discussed with olofj, and we think the right thing is to use a waitqueue
and wait for there to be work available or for the thread to be asked to exit.

BUG=chromium-os:14205
TEST=Adhoc
Build, modem status, suspend/resume

http://codereview.chromium.org/6913008/