msm_serial: Don't configure serial interrupt as wakeup source in probe
Serial device interrupt is configured as wakeup source in probe. Due to
this, even there is no serial device user on APPS side, and same serial
instance is getting used on Modem. Serial interrupt is received by APPS
as well which is waking up the APPS. Hence configure serial interrupt as
wakeup source in startup to make sure that it is configured only when there
is user for the same.

Change-Id: Idb51ad28ffadee08e659ca0816e3fa3e2d3b0297
CRs-fixed: 284567
Signed-off-by: Mayank Rana <mrana@codeaurora.org>