CHROMIUM: USB: EHCI: fix random suspend failures
When a resume from autosuspend is done (ie the reset_done counter has
expired) but the resume completed has not been detected yet (reset_done
counter not zeroed), we should not cancel the explicit suspend requests.
s the resume is not longer on-going.

BUG=chrome-os-partner:3076
TEST=manually run series of "powerd_suspend" with AC unplugged and check that
the board is suspending every time.
(the power_Resume autotest cannot be used here, since it would require
to have the USB-ethernet token plugged which interferes with the USB hub
suspend)

Change-Id: If95000f3bf6b1e101880a62d257032a0174365de
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/337
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/540