video: tegra: dc: nvhdcp diagnostics
Add a diagnostics function for nvhdcp to test and verify the a and d ksv's
without the need to attach and hdmi device. This path is intended to be
used carefully, and only verifies the validity of the ksvs.

This adds a debug fs entry and can be accessed via:

cat /sys/kernel/debug/nvhdcp/diagnostics

TEST=once booted, can successfully read ksv's
BUG=chrome-os-partner:3164

Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Cherry picked from Change-Id: Id96ea13e31627263c15601d6bdcfd1d783466b1b
Reviewed-on: http://gerrit.chromium.org/gerrit/839
Reviewed-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>

Change-Id: I230132bbd28844a748aa27e79eae62187e639046
Reviewed-on: http://gerrit.chromium.org/gerrit/1562
Reviewed-by: Micah Catlin <micahc@chromium.org>
Tested-by: Micah Catlin <micahc@chromium.org>