CHROMIUM: gobi: Don't kill urbs in failed synchronous writes.
If we kill urbs, we may or may not send what we meant to send, which is fine in
most cases, since we're about to report the failure to userspace anyway.
However, if the connection manager is being killed, the driver needs to tear
down its remaining QMI connections itself. Furthermore, if we're being killed,
we can't necessarily down_interruptible(), since we have a pending signal, so we
want to send those urbs on to the card whether or not down_interruptible()
succeeds.

In the 'common' case (i.e., no interrupt), this change makes no functional
difference. If we receive a signal, it guarantees that the urb we submitted is
sent, where previously whether it had been sent or not was a matter of chance.

TEST=network_3GStressEnable, network_GobiUncleanDisconnect
BUG=chromium-os:14945

Change-Id: I1cc558dd748b0db11edfd592c8caff832d97ea46
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/431
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>