msm: rpc: Split SDIO_XPRT buffers equally for Tx and Rx.
During the system bringup, when there is no RPC Router flow control
in action, all the SDIO_XPRT buffers are consumed by incoming messages.
This leads to non-availability of buffers for the response messages
to these incoming messages. Hence RPC router virtually enters into a
deadlock. So split the buffers equally for Tx & Rx so that no single
activity consumes the entire buffers. If there are no buffers available
to consume the incoming messages, complete notifying the RPC Router if
there is data to be processed by it.

Change-Id: Ic373ad2c327dc5b361caa0b3564f4f76521d016f
CRs-Fixed: 286478, 294176
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>