CHROMIUM: Refactor chromeos platform layer to cover both arm and x86.
This CL completes the set of modifications required to support
chromeos ARM platforms (see
http://codereview.chromium.org/6883204 and
http://codereview.chromium.org/6902164 submitted earlier).

The chromeos.c interface layer is refactored: the platform
independent components consolidated, and platform specific ones
are moved to the appropriate platform modules. The ability to run
arm/chromos_arm.c as a module is eliminated, and the default
value of the shared memory window base is set.

The unused APIs (chromeos_is_devmode() and
chromeos_initialized()) are removed, appropriate Kconfig and
Makefile changes are made to reflect single chromeos.c module for
both platforms.

fs/proc/breakme.c is modified to allow requesting a recovery mode
reboot, to facilitate testing of the chromeos API.

BUG=chromium-os:12522
TEST=manual:

. build the new kernel image for mario, alex and kaen

. on all platforms run `crossystem' and observe the results (the
  results look reasonable).

. on all systems request reboot in recovery mode using `crossystem
  recovery_request=0xc1' and reboot. Observe the systems come back in
  recovery mode.

. on mario and alex recuest recovery by executing `echo need_recovery
  > /proc/breakme' and reboot. Observe the systems come back in
  recovery mode.

Change-Id: Ifa30aafa110479ee8981325a5e08fb3127ef8b6e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/450
Reviewed-by: Olof Johansson <olofj@chromium.org>
[olofj: removed breakme modification when splitting branches]
Signed-off-by: Olof Johansson <olofj@chromium.org>