UPSTREAM: ASoC: WM8903: Fix Digital Capture Volume range
Increase the range of the Digital Capture Volume control to be 120 steps.
Each step is 0.75dB, and the range starts at -72dB, giving a max setting
of 18dB, which matches the latest datasheet, to the precision of the step
size.

Cherrypicked from Change-Id: I10040617dc160cc1ad131bad0077f3b50bc3179d
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Acked-by: Liam Girdwood <lrg@ti.com>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/600
Reviewed-by: Olof Johansson <olofj@chromium.org>

Change-Id: I1310a5de95adce5b6751ec7006d857baead1b25a
Reviewed-on: http://gerrit.chromium.org/gerrit/994
Reviewed-by: Micah Catlin <micahc%chromium.org@gtempaccount.com>
Tested-by: Micah Catlin <micahc%chromium.org@gtempaccount.com>