msm: clock-local: Replace readl() and writel() with _relaxed versions
The local clock driver makes an effort to ensure proper barriers
are inserted between register accesses that are in different 1K
alignments when ordering matters. For this reason, it is safe to
use the _relaxed versions of readl() and write() to avoid any
performance loss caused by the unnecessary memory barriers inside
those macros.

Change-Id: Iaad267da0016effa5f5c3e255562f3ecc573be34
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>