msm: devices-7x30: Remove support for pbus_clk.
PBUS_CLK is no longer needed since we have a separate
EBI1_FIXED_CLK that remains effective through power-collase.

Since PBUS_CLK was only being used as a hack to influence the
EBI1 clock during power-collapse, and this is now handled by
EBI1_FIXED_CLK, we can remove the unused clock and the code
needed to support it.

Change-Id: Ifb54dc61f2af8942dff15c2d86b757171a4d6b90
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>