msm: adsp: Update with H264 and VP8 codec changes.
This change is to support H264 and VP8 codecs. 0xB201 is the frameheader
ID for H264 and 0xFD01 is the frame header ID for VP8 codec slice data.
Slice data has YUV buffer addresses where DSP has to write decoded data,
but these addresses are virtualaddresses. This code is added to convert
virtual addresses to physical addresses before passing them to DSP

Change-Id: I2272f1bc7a66a88fcc068adf7997e2ae8be0f658
Signed-off-by: mlanka <mlanka@codeaurora.org>