CHROMIUMOS: mac80211: Start monitor work on restart
Trigger connection monitor on resume from suspend.  Since we
have been sleeping, there is reason to suspect that we might
not still be associated.  The speed of detecting loss of
{connection,authentication} is worth the cost of the small
additional traffic at resume.

Sumbitter note -- I've tested this for regressions, but
I need reviewer to test under conditions that originally
caused the bug, as I have been unsuccessful in replicating
them.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chromium-os:14037
TEST=Redo Roaming tests

Change-Id: Idec7ec4a834c9dd2745aa9b96212460a5899c40c
Reviewed-on: http://gerrit.chromium.org/gerrit/1221
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>