msm: board-msm8960: add PMIC 8921 GPIO and MPP initial configs
The configurations of PMIC GPIOs and MPPs are board specific and
must be performed after the PMIC GPIO and MPP drivers have probed,
but before any clients modify the pins via gpiolib functions.  Add
a temporary function which configures some of the GPIOs and MPPs
in the board file.  This will be achieved differently once some
kind of pinmux is available.

Change-Id: I5437614940e6413c05160224bffea82d354e2b89
Signed-off-by: David Collins <collinsd@codeaurora.org>