msm: board-7x27a: Use correct enums for USB host VBUS gpio
SURF and FFA are having different enums eventhough these are
using the same gpio expander. Hence use the correct enum for
USB host VBUS gpio based on the machine macro.

OTG driver takes care of turning on/off of the VBUS. Hence VBUS
power hook is not required in the host platform data.

Change-Id: I1db53de351f73b0cdda66d3f7d57fe0c9754bc6b
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>