misc: pmic8058-xoadc: Add mpp_config call on configure
Call mpp_config during xoadc read request. This is done
to configure the mpp mapping on scenario's when the remote
processor overwrites the configuration. The corner case for
this scenario is if a overwrite to the MPP occures in the
middle of the ADC transaction. By design, the MPP mapping
is configured to be the same between the ADC drivers.

CRs-Fixed: 286523
Change-Id: I173c50ba59488be137b198eada1f4f2d8ef4a01d
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>