defconfig: msm7627a: Disable YAFFS from doing ECC on tags by default
YAFFS2 require 28bytes of OOB area when it does ECC on tags.
But for a 2K page NAND device only 16bytes of spare area is
reserved for YAFFS tags. Which can cause mount failures and
loss of data after every remount operation. Hence, disable
YAFFS from doing ECC on tags as the ECC layout of underlying
NAND driver is taking care of computing hardware ECC on the
OOB spare area.

Change-Id: Idb0560062156bf2ba0b1e8b518b8065dba832e90
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>