FACTORY: CHROMIUM: video: tegra: Remove fine-grained DDC hooks
It turns out that some HDMI sinks actually need +5V all the time, since
they rely on it to provide the hotplug-detect current back to the
source.  So we can't turn it on only when doing DDC or TMDS traffic, but
need to leave it on all the time.

A subsequent change will add hooks to enable and disable based on
suspend state.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

BUG=cros-partner:1812
TEST=none
Cherry picked from Change-Id: Id464f1394fd2c28acaa29efdf1507aa47980fc44
Reviewed-on: http://gerrit.chromium.org/gerrit/1093
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>

Change-Id: Ife5c72ff630b66c98404cacc6faa16083f86bd48
Reviewed-on: http://gerrit.chromium.org/gerrit/1566
Tested-by: Micah Catlin <micahc@chromium.org>
Reviewed-by: Micah Catlin <micahc@chromium.org>