USB: gadget: msm72k_udc: Replace dma_coherent_pre/post_ops with mb/rmb
The current implementation of dma_coherent_pre_ops uses dmb() which is
not sufficient. The DMB only ensures the ordering of accesses with
regards to processors accessing the same memory. If a DMA transfer is
started by a write to Device memory, the data to be transfered may not
reach the main memory (even if mapped as Normal Non-cacheable) before the
device receives the notification to begin the transfer. The only barrier
that would help in this situation is DSB which would completely drain the
write buffers.

mb()/rmb()/wmb() implementation available which takes care of different
barrier requirement based on different platform and provide fool-proof
barriers to use by device driver. Hence use mb() and rmb() instead of
dma_coherent_pre_ops() and dma_coherent_post_ops().

Change-Id: I82413e0e345b41b26a261ac8f34dbec3ec1cc777
Signed-off-by: Mayank Rana <mrana@codeaurora.org>