mfd-core: WCD9310: Toggle codec reset line on bootup
The WCD9310 reset line must be toggled on bootup to
reset the internal state of the codec. This reset must
occur before the slim_get_logical_addr function is
called, or the function call will fail.

The 20 ms delay between each stage of toggling the reset
line ensures that the codec is fully reset before its
registers are accessed and it gives the codec hardware
enough time to enumerate with SLIMBUS. Twenty ms is
enough time to meet the codec's hardware specification,
and the duration of the delay may be able to be
decreased after more testing.

Change-Id: I5488976ffaa00d152a11614cd974d9201d46cc1d
Signed-off-by: Brad Rubin <brubin@codeaurora.org>